Fix output writer for doctrine-migrations v2 integration
When we use doctrine-migrations v2 we can't see output on the console
on running commands, for example:
```
vendor/bin/doctrine-module migrations:migrate
```

This is because in v2 we do have dependency factory and the same instance
of dependency is always used instead of creating new.
In the migration configuration factory however we are creating another
instance which is not used in dependency factory later on.