Fix too restrictive return type for `getItems()` of `DoctrinePaginator`
The Laminas `AdapterInterface` is absolutely not restrictive of the type
of the items that can be in the paginator's collection. However, here
the incorrect Psalm type was used.

The `DoctrinePaginator` expects an `object` (typed as `T`), such as
`Doctrine\ORM\Tools\Pagination\Paginator`. However, this `T` is not the
correct type for the items in the paginator.